refactoring out models
